requirements.txt
streamlit
panda
numpy
plotly.express
openpyxl
